SDCP-306/7: Improved handling of user-provided cell and symmetry info
Allow workflow user to provide spacegroup/pointgroup and optionally unit
cell parameters to the initial indexing of the characterisation data
collection. This is because the table of indexing solutions produced by
IDXREF sometimes doesn't include the correct one.

Also allow user to indicate to aP_wf_process that the provided cell and
symmetry should be made use of